requirements.txt
numpy~=1.24.0
matplotlib~=3.6.2
seaborn~=0.12.1
spacy~=3.4.4
streamlit~=1.16.0
pillow~=9.3.0
pandas~=1.5.2
pymongo~=3.12.0
joblib~=1.2.0
plotly~=5.11.0
setuptools~=65.5.0
contractions~=0.0.52
emoji~=1.6.1
scikit-learn~=1.2.0